:hammer: refactor: 데이터 코드 분리 및 반환값 변경
